+++ Tests/run.js
@@ -1,3 +1,4 @@
 require("./MeituanBikeWeChatAds.test.js");
 require("./XiaohongshuHomeFeedAds.test.js");
 require("./MeituanAppAdMedia.test.js");
+require("./WeChatMiniGameAdTelemetry.test.js");